requirements.txt
numpy>=1.21.0
torch>=1.10.0
stable-baselines3>=1.6.0
gym>=0.26.0
tensorflow>=2.7.0
pytest>=7.0.0
python-dotenv>=0.19.0